--- src/lib/handy_funcs.js
@@ -1,28 +1,0 @@
-const flatten = (a, b) => {
-   return a.concat(Array.isArray(b) ? b.reduce(flatten, []) : b);
-}
-
-function flattenIteration(arr, flatArr) {
-   flatArr = flatArr || [];
-
-   var length = arr.length | 0;
-
-   for (var index = 0; index < length; index = index + 1) {
-      var item = arr[index];
-      item.constructor === Array ? flattenIteration(item, flatArr) : flatArr[flatArr.length] = item;
-   }
-
-   return flatArr;
-}
-module.exports = {
-   smoothArray: (nested) => {
-      if (!nested) return []
-      return nested.reduce(flatten, []).filter((ne) => ne !== null && ne !== undefined);
-   },
-   flatten: (nested) => {
-      return nested.reduce(flatten, []);
-   },
-   capitalize: (string) => {
-      return string.charAt(0).toUpperCase() + string.slice(1);
-   }
-}
