+++ src/lib/utils/form_tags.js
@@ -1,0 +1,9 @@
+const formTags = {
+   textarea: true,
+   select: true,
+   input: true,
+   output: true,
+   form: true
+}
+
+export default formTags
