+++ src/lib/utils/svg_tags.js
