+++ webpack.config.js
@@ -4,6 +4,15 @@ const path = require('path');
 module.exports = {
   devtool: debug ? "inline-sourcemap" : false,
   entry: "./src/ex.js",
+  resolve: {
+   extensions: ['.js', ''],
+   alias: {
+     addon: path.resolve( __dirname, 'src/addons'),
+     lib: path.resolve( __dirname, 'src/lib'),
+     shared: path.resolve( __dirname, 'src/shared'),
+     utils: path.resolve( __dirname, 'src/lib/utils')
+   }
+ },
   module: {
     loaders: [
       {
